feat(cart): replace cart store with transaction model
Promote the transaction prototype to the production cart store so optimistic updates, cancellation, settlement, and errors have explicit keyed ownership.

Handle asynchronous Standard Events and both endpoint response shapes, synchronize uncontrolled quantity inputs, preserve visible pending states, and include focused regression coverage plus a patch changeset.
