+++ .changeset/cart-transaction-model.md
@@ -1,0 +1,5 @@
+---
+"@shopify/hydrogen": patch
+---
+
+Replace cart optimistic update internals with keyed transaction and error projections, including reliable cancellation, request settlement, and pending state updates.
