+++ .oxlintrc.json
@@ -117,7 +117,7 @@
         "packages/hydrogen/src/core/predictive-search/server-handlers.ts",
         "packages/hydrogen/src/core/predictive-search/store.ts",
         "packages/hydrogen/src/core/cart/standard-actions-adapter.ts",
-        "packages/hydrogen/src/core/cart/transaction-cart-store.ts",
+        "packages/hydrogen/src/core/cart/cart.ts",
         "packages/hydrogen/src/core/cart/state.ts",
         "packages/hydrogen/src/core/collection/collection.ts",
         "packages/hydrogen/src/core/interceptors/graphiql.ts",
@@ -159,7 +159,6 @@
         "packages/hydrogen/src/core/shopify-scripts/consent.ts",
         "packages/hydrogen/src/core/shopify-scripts/utils/tracking-values.ts",
         "packages/hydrogen/src/core/cart/actions.ts",
-        "packages/hydrogen/src/core/cart/cart.ts",
         "packages/hydrogen/src/core/cart/queries.ts",
         "packages/hydrogen/src/core/predictive-search/queries.ts",
         "packages/hydrogen/src/core/predictive-search/search.ts",
